+++ source/ItemHeights.js
@@ -1,25 +1,28 @@
 import log from './log'
 
 export default class ItemHeights {
-	constructor(getContainerNode, getState) {
-		this.getContainerNode = getContainerNode
+	constructor(getContainerElement, getState) {
+		this.getContainerElement = getContainerElement
 		this.getState = getState
+		this.initialize()
+	}
+
+	initialize() {
 		this.reset()
+		if (this.getState()) {
+			this.onStateUpdate()
+		}
 	}
 
 	reset() {
 		this.measuredItemsHeight = 0
 		this.firstMeasuredItemIndex = undefined
 		this.lastMeasuredItemIndex = undefined
-		// this.averageItemHeight = undefined
-		// this.averageItemHeightSamplesCount = undefined
-		// this.previousAverageItemHeight = undefined
-		// this.previousAverageItemHeightSamplesCount = undefined
 	}
 
 	/**
 	 * Initializes `this.measuredItemsHeight`, `this.firstMeasuredItemIndex` and
-	 * `this.lastMeasuredItemIndex` instance variables.
+	 * `this.lastMeasuredItemIndex` instance variables from the `state`.
 	 * These instance variables are used when calculating "average" item height:
 	 * the "average" item height is simply `this.measuredItemsHeight` divided by
 	 * `this.lastMeasuredItemIndex` minus `this.firstMeasuredItemIndex` plus 1.
@@ -28,11 +31,10 @@ export default class ItemHeights {
 	 * jumps from one position to a distant another position. How could that happen?
 	 * Maybe it can't, but just in case.
 	 */
-	onInitItemHeights() {
-		this.reset()
+	onStateUpdate() {
 		let i = 0
 		while (i < this.getState().itemHeights.length) {
-			if (this.getState().itemHeights[i] == undefined) {
+			if (this.getState().itemHeights[i] === undefined) {
 				if (this.firstMeasuredItemIndex !== undefined) {
 					this.lastMeasuredItemIndex = i - 1
 					break
@@ -61,7 +63,7 @@ export default class ItemHeights {
 	// }
 
 	_getItemHeight(i, firstShownItemIndex) {
-		const container = this.getContainerNode()
+		const container = this.getContainerElement()
 		if (container) {
 			const nodeIndex = i - firstShownItemIndex
 			if (nodeIndex >= 0 && nodeIndex < container.childNodes.length) {
@@ -73,7 +75,7 @@ export default class ItemHeights {
 	}
 
 	getItemSpacing() {
-		const container = this.getContainerNode()
+		const container = this.getContainerElement()
 		if (container) {
 			if (container.childNodes.length > 1) {
 				const firstItem = container.childNodes[0]